requirements.txt
fastapi
uvicorn[standard]
jinja2
rubik-cube